+++ .travis.yml
@@ -6,4 +6,4 @@ install:
   - pip install -r requirements.txt
 
 script:
-  - pylint delete_files.py workers create_messages.py rabbitmq_setup.py
+  - pylint export_workers
