+++ export_workers/workers/create_file/serializers/form_schema.py
@@ -1,0 +1,9 @@
+from marshmallow import Schema, fields
+
+class FormResponseSchema(Schema):
+    """Schema for response from form service"""
+    description = fields.String(required=True)
+    title = fields.String(required=True)
+    form_id = fields.Integer(required=True)
+    owner = fields.String(required=True)
+    fields = fields.List(fields.Integer(), required=True)
